Implement index tracking in SortedSet
This permits sorted set content changes to provide an index argument,
suitable for binding changes to an array, or an array that provides a
window into a portion of a sorted set.

It also permits a trivial implementation of indexOf with amortized
logarithmic average speed relative to the length of the set.

This involves incrementally tracking the number of nodes in every
subtree, which necessitates a twist to the top-down splaying algorithm.
Every root node that is bypassed while linking left and right gets
tracked in a history tree.  The wings of that tree are visited after the
reassembly to propagate the new subtree lengths up to the new root.