Urgent fix to map construction
Map constructors depend on a way to distinguish an array of duples from
a map.  Up to this point, arrays have not had keys, values, or entries
methods, intrisinc to map types and this was used to distinguish them
from arrays of duples.  ES6 proposes keys, values, and entries methods
for arrays (which serve as index to value maps) and these have been
implemented on Chrome 30.

Since there are no longer any methods of a map that are not also on
Array, it has become necessary for map-like collections to advertise an
`isMap` property of `true`.