Make SortedSet `one` consistent between changes.
The sorted set previously returned the root node's value for speed.
Because of this, `one` would return different values based on
the most recent node *accessed*, which means that it would change
between change notifications.  This would interfere with the
consistency requirements of bindings.  At the expense of requiring
logarithmic time, `min` produces a consistent value between
changes.