Various swap and splice fixes
This change reimplements swap in plain JavaScript so that it can avoid
the problems with splice, including the large array problem previously
solved, preserves holes in the plus array, and extends the array if the
starting position is beyond the length.

Splice truncates the start index to the length. As such, splice is not a
suitable primitive for the set method, since setting the value at an
index should extend the array and leave holes behind. Splice is also not
good at copying array holes.

Fix array set and swap for indexes beyond the initial length.

:warning: Additionally, I have removed the behavior of returning the
array of removed values. This is seldom useful and usually generates
unnecessary garbage.

Fix redundant property change on array indexes

Add failing test for redundant property dispatch on unchanged properties
of an array during swap.

Add test verifying that map changes only dispatch for value change on
update.

Fix problem with observing the last index of an array. Off by one error
in the book keeping logic.

Reconcile observed and unobserved array swap.