Ignore property changes on non-configurables
In Safari, properties of hosted objects (particularly DOM elements) are
not configurable so observing a property on a DOM element, even if to
traverse to an observable object, would fail with an error. This change
allows observers to proceed on the assumption that the non-configurable
property will not be changed.