Fix dispatchesMapChanges check in Dict
Every other place in the source that a class checks if it should dispatch
changes with this.dispatchBeforeMapChange looks to
this.dispatchMapChanges.  In Dict.set, it looks to
this.dispatchesBeforeMapChanges, which is the only line in the whole
codebase that references this variable.  Changing to dispatchesMapChanges
as I think it was supposed to be.

Signed-off-by: Fred Kilbourn <fred@fredk.com>