Add flag to allow active descriptors in listeners
There are times when listeners might want to make changes that trigger
further events in a collection.  I had a bug in my code around this
point and only realized that further collection changes are suppressed
by default.  This patch adds an override flag to allow for descriptors
to fire their listeners again even if currently active.

Signed-off-by: Fred Kilbourn <fred@fredk.com>